--- src/app/components/ParticlesBackground.tsx
@@ -1,125 +1,0 @@
-// src/components/ParticlesBackground.tsx
-
-"use client";
-
-import React from "react";
-import { loadFull } from "tsparticles";
-import { Engine } from "tsparticles-engine";
-import { Particles } from "react-tsparticles";
-
-const ParticlesBackground: React.FC = () => {
-  const particlesInit = async (main: Engine) => {
-    await loadFull(main);
-  };
-
-  return (
-    <div className="absolute inset-0 -z-10">
-      <Particles
-        id="tsparticles"
-        init={particlesInit}
-        options={{
-          fullScreen: {
-            enable: false,
-            zIndex: -1,
-          },
-          particles: {
-            number: {
-              value: 380,
-              density: {
-                enable: true,
-                area: 800,
-              },
-            },
-            color: {
-              value: "#ffffff",
-            },
-            shape: {
-              type: "circle",
-              stroke: {
-                width: 0,
-                color: "#000000",
-              },
-              polygon: {
-                nb_sides: 5,
-              },
-              image: {
-                src: "img/github.svg",
-                width: 100,
-                height: 100,
-              },
-            },
-            opacity: {
-              value: 0.5,
-              anim: {
-                enable: false,
-              },
-            },
-            size: {
-              value: 3,
-              random: true,
-              anim: {
-                enable: false,
-              },
-            },
-            links: {
-              enable: true,
-              distance: 150,
-              color: "#ffffff",
-              opacity: 0.4,
-              width: 1,
-            },
-            move: {
-              enable: true,
-              speed: 6,
-              direction: "none",
-              outModes: {
-                default: "out",
-              },
-            },
-          },
-          interactivity: {
-            events: {
-              onHover: {
-                enable: true,
-                mode: "grab",
-              },
-              onClick: {
-                enable: true,
-                mode: "push",
-              },
-              resize: true,
-            },
-            modes: {
-              grab: {
-                distance: 140,
-                links: {
-                  opacity: 1,
-                },
-              },
-              bubble: {
-                distance: 400,
-                size: 40,
-                duration: 2,
-                opacity: 8,
-                speed: 3,
-              },
-              repulse: {
-                distance: 200,
-                duration: 0.4,
-              },
-              push: {
-                quantity: 4,
-              },
-              remove: {
-                quantity: 2,
-              },
-            },
-          },
-          retina_detect: true,
-        }}
-      />
-    </div>
-  );
-};
-
-export default ParticlesBackground;
