+++ src/app/page.tsx
@@ -4,27 +4,9 @@
 
 import Image from "next/image";
 import ParticlesBackground from "@/components/ParticlesBackground";
-import { useEffect, useState } from "react";
+import DateTime from "@/components/DateTime"; // Import the new component
 
 export default function Home() {
-  const [dateTime, setDateTime] = useState<string>("");
-
-  useEffect(() => {
-    const updateDateTime = () => {
-      const now = new Date();
-      const formatted = now.toLocaleString("en-US", {
-        dateStyle: "full",
-        timeStyle: "medium",
-      });
-      setDateTime(formatted);
-    };
-
-    updateDateTime();
-    const interval = setInterval(updateDateTime, 20000);
-
-    return () => clearInterval(interval);
-  }, []);
-
   return (
     <div className="relative min-h-screen overflow-hidden bg-black text-white">
       {/* Particle Background */}
@@ -42,9 +24,7 @@ export default function Home() {
           />
 
           {/* Date and Time in Yellow */}
-          <div className="text-yellow-500 text-lg font-semibold">
-            {dateTime}
-          </div>
+          <DateTime /> {/* Use the new DateTime component here */}
 
           <ol className="list-inside list-decimal text-sm text-center sm:text-left font-[family-name:var(--font-geist-mono)]">
             <li className="mb-2">
@@ -56,80 +36,7 @@ export default function Home() {
             </li>
             <li>Save and see your changes instantly. Nextjs!!!</li>
           </ol>
-
-          <div className="flex gap-4 items-center flex-col sm:flex-row">
-            <a
-              className="rounded-full border border-solid border-transparent transition-colors flex items-center justify-center bg-foreground text-background gap-2 hover:bg-[#383838] dark:hover:bg-[#ccc] text-sm sm:text-base h-10 sm:h-12 px-4 sm:px-5"
-              href="https://vercel.com/new"
-              target="_blank"
-              rel="noopener noreferrer"
-            >
-              <Image
-                className="dark:invert"
-                src="/vercel.svg"
-                alt="Vercel logomark"
-                width={20}
-                height={20}
-              />
-              Deploy now
-            </a>
-            <a
-              className="rounded-full border border-solid border-black/[.08] dark:border-white/[.145] transition-colors flex items-center justify-center hover:bg-[#f2f2f2] dark:hover:bg-[#1a1a1a] hover:border-transparent text-sm sm:text-base h-10 sm:h-12 px-4 sm:px-5 sm:min-w-44"
-              href="https://nextjs.org/docs"
-              target="_blank"
-              rel="noopener noreferrer"
-            >
-              Read our docs
-            </a>
-          </div>
         </main>
-        <footer className="row-start-3 flex gap-6 flex-wrap items-center justify-center">
-          <a
-            className="flex items-center gap-2 hover:underline hover:underline-offset-4"
-            href="https://nextjs.org/learn"
-            target="_blank"
-            rel="noopener noreferrer"
-          >
-            <Image
-              aria-hidden
-              src="/file.svg"
-              alt="File icon"
-              width={16}
-              height={16}
-            />
-            Learn
-          </a>
-          <a
-            className="flex items-center gap-2 hover:underline hover:underline-offset-4"
-            href="https://vercel.com/templates?framework=next.js"
-            target="_blank"
-            rel="noopener noreferrer"
-          >
-            <Image
-              aria-hidden
-              src="/window.svg"
-              alt="Window icon"
-              width={16}
-              height={16}
-            />
-            Examples
-          </a>
-          <a
-            className="flex items-center gap-2 hover:underline hover:underline-offset-4"
-            href="https://nextjs.org"
-            target="_blank"
-            rel="noopener noreferrer"
-          >
-            <Image
-              aria-hidden
-              src="/globe.svg"
-              alt="Globe icon"
-              width={16}
-              height={16}
-            />
-            Go to nextjs.org → Official website
-          </a>
-        </footer>
       </div>
     </div>
   );
